# workspaces/litellm-orchestrator/plugins/litellm-orchestrator-backend/src/schema/openapi.yaml
openapi: 3.1.0
info:
  title: litellm-orchestrator
  version: 1.0.0
tags:
  - name: users
    description: User management
paths:
  /users:
    post:
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
              required:
                - userId
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}:
    get:
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/keys:
    post:
      operationId: createUserKey
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: User key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserKey'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: getUserKeys
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User keys found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserKeyList'
        '404':
          description: User keys not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/keys/{keyId}:
    delete:
      operationId: deleteUserKey
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User key deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        name:
          type: string
        message:
          type: string
    User:
      type: object
      properties:
        userId:
          type: string
        createdAt:
          type: string
        maxBudget:
          type: number
        budgetDuration:
          type: string
        spend:
          type: number
      required:
        - userId
        - createdAt
        - maxBudget
        - budgetDuration
        - spend
      title: User
      description: A user in the system
    UserKeyList:
      type: object
      properties:
        keys:
          type: array
          items:
            $ref: '#/components/schemas/UserAPIKeyAuth'
      required:
        - keys
      title: User keys
      description: A user's keys
    UserAPIKeyAuth:
      properties:
        token:
          anyOf:
            - type: string
            - type: 'null'
          title: Token
        key_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Key Name
        key_alias:
          anyOf:
            - type: string
            - type: 'null'
          title: Key Alias
        spend:
          type: number
          title: Spend
          default: 0
        max_budget:
          anyOf:
            - type: number
            - type: 'null'
          title: Max Budget
        expires:
          anyOf:
            - type: string
            - type: string
              format: date-time
            - type: 'null'
          title: Expires
        models:
          items: {}
          type: array
          title: Models
          default: []
        aliases:
          additionalProperties: true
          type: object
          title: Aliases
          default: {}
        config:
          additionalProperties: true
          type: object
          title: Config
          default: {}
        user_id:
          anyOf:
            - type: string
            - type: 'null'
          title: User Id
        team_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Team Id
        max_parallel_requests:
          anyOf:
            - type: integer
            - type: 'null'
          title: Max Parallel Requests
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
          default: {}
        tpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: Tpm Limit
        rpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: Rpm Limit
        budget_duration:
          anyOf:
            - type: string
            - type: 'null'
          title: Budget Duration
        budget_reset_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Budget Reset At
        allowed_cache_controls:
          anyOf:
            - items: {}
              type: array
            - type: 'null'
          title: Allowed Cache Controls
          default: []
        allowed_routes:
          anyOf:
            - items: {}
              type: array
            - type: 'null'
          title: Allowed Routes
          default: []
        permissions:
          additionalProperties: true
          type: object
          title: Permissions
          default: {}
        model_spend:
          additionalProperties: true
          type: object
          title: Model Spend
          default: {}
        model_max_budget:
          additionalProperties: true
          type: object
          title: Model Max Budget
          default: {}
        soft_budget_cooldown:
          type: boolean
          title: Soft Budget Cooldown
          default: false
        blocked:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Blocked
        litellm_budget_table:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Litellm Budget Table
        org_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Org Id
        created_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Created At
        created_by:
          anyOf:
            - type: string
            - type: 'null'
          title: Created By
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
        updated_by:
          anyOf:
            - type: string
            - type: 'null'
          title: Updated By
        object_permission_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Object Permission Id
        object_permission:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
        rotation_count:
          anyOf:
            - type: integer
            - type: 'null'
          title: Rotation Count
          default: 0
        auto_rotate:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Auto Rotate
          default: false
        rotation_interval:
          anyOf:
            - type: string
            - type: 'null'
          title: Rotation Interval
        last_rotation_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Last Rotation At
        key_rotation_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Key Rotation At
        team_spend:
          anyOf:
            - type: number
            - type: 'null'
          title: Team Spend
        team_alias:
          anyOf:
            - type: string
            - type: 'null'
          title: Team Alias
        team_tpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: Team Tpm Limit
        team_rpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: Team Rpm Limit
        team_max_budget:
          anyOf:
            - type: number
            - type: 'null'
          title: Team Max Budget
        team_models:
          items: {}
          type: array
          title: Team Models
          default: []
        team_blocked:
          type: boolean
          title: Team Blocked
          default: false
        soft_budget:
          anyOf:
            - type: number
            - type: 'null'
          title: Soft Budget
        team_model_aliases:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Team Model Aliases
        team_member:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
        team_metadata:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Team Metadata
        team_object_permission_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Team Object Permission Id
        team_member_spend:
          anyOf:
            - type: number
            - type: 'null'
          title: Team Member Spend
        team_member_tpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: Team Member Tpm Limit
        team_member_rpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: Team Member Rpm Limit
        end_user_id:
          anyOf:
            - type: string
            - type: 'null'
          title: End User Id
        end_user_tpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: End User Tpm Limit
        end_user_rpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: End User Rpm Limit
        end_user_max_budget:
          anyOf:
            - type: number
            - type: 'null'
          title: End User Max Budget
        organization_max_budget:
          anyOf:
            - type: number
            - type: 'null'
          title: Organization Max Budget
        organization_tpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: Organization Tpm Limit
        organization_rpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: Organization Rpm Limit
        organization_metadata:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Organization Metadata
        last_refreshed_at:
          anyOf:
            - type: number
            - type: 'null'
          title: Last Refreshed At
        api_key:
          anyOf:
            - type: string
            - type: 'null'
          title: Api Key
        user_role:
          anyOf:
            - type: string
            - type: 'null'
        allowed_model_region:
          anyOf:
            - type: string
              enum:
                - eu
                - us
            - type: 'null'
          title: Allowed Model Region
        parent_otel_span:
          anyOf:
            - {}
            - type: 'null'
          title: Parent Otel Span
        rpm_limit_per_model:
          anyOf:
            - additionalProperties:
                type: integer
              type: object
            - type: 'null'
          title: Rpm Limit Per Model
        tpm_limit_per_model:
          anyOf:
            - additionalProperties:
                type: integer
              type: object
            - type: 'null'
          title: Tpm Limit Per Model
        user_tpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: User Tpm Limit
        user_rpm_limit:
          anyOf:
            - type: integer
            - type: 'null'
          title: User Rpm Limit
        user_email:
          anyOf:
            - type: string
            - type: 'null'
          title: User Email
        request_route:
          anyOf:
            - type: string
            - type: 'null'
          title: Request Route
      type: object
      title: UserAPIKeyAuth
      description: Return the row in the db
    UserKey:
      type: object
      properties:
        userId:
          type: string
        keyId:
          type: string
        keySecret:
          type: string
      required:
        - userId
        - keyId
        - keySecret
      title: User key
      description: A user key